+++ crates/mysten-common/src/lib.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod metrics;
 pub mod sync;
