+++ crates/sui-bridge/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 
 [dependencies]
 ethers = "2.0"
-snap = "1.1.0"
 tokio = { workspace = true, features = ["full"] }
 sui-types.workspace = true
 sui-authority-aggregation.workspace = true
@@ -26,7 +25,6 @@ mysten-metrics.workspace = true
 sui-sdk.workspace = true
 sui-keys.workspace = true
 sui-config.workspace = true
-sui-tls.workspace = true
 clap.workspace = true
 tracing.workspace = true
 bin-version.workspace = true
@@ -46,6 +44,7 @@ rand.workspace = true
 lru.workspace = true
 shared-crypto.workspace = true
 backoff.workspace = true
+mysten-common.workspace = true
 enum_dispatch.workspace = true
 sui-json-rpc-api.workspace = true
 sui-test-transaction-builder.workspace = true
